Move `ChainEndpoint` trait into its own module
