+++ relayer-cli/src/cli_utils.rs
@@ -3,12 +3,12 @@ use tokio::runtime::Runtime as TokioRuntime;
 
 use ibc::core::ics02_client::client_state::ClientState;
 use ibc::core::ics24_host::identifier::{ChainId, ChannelId, PortId};
-use ibc_relayer::chain::counterparty::{channel_connection_client, ChannelConnectionClient};
 use ibc_relayer::{
     chain::{
+        cosmos::CosmosSdkChain,
+        counterparty::{channel_connection_client, ChannelConnectionClient},
         handle::{BaseChainHandle, ChainHandle},
         runtime::ChainRuntime,
-        CosmosSdkChain,
     },
     config::Config,
 };
