+++ relayer-cli/src/commands/query/connections.rs
@@ -6,7 +6,8 @@ use ibc_relayer::chain::requests::{PageRequest, QueryConnectionsRequest};
 use tokio::runtime::Runtime as TokioRuntime;
 
 use ibc::core::ics24_host::identifier::{ChainId, ConnectionId};
-use ibc_relayer::chain::{ChainEndpoint, CosmosSdkChain};
+use ibc_relayer::chain::cosmos::CosmosSdkChain;
+use ibc_relayer::chain::endpoint::ChainEndpoint;
 
 use crate::conclude::{exit_with_unrecoverable_error, Output};
 use crate::prelude::*;
