+++ relayer/src/transfer.rs
@@ -13,9 +13,9 @@ use ibc::Height;
 use ibc_proto::google::protobuf::Any;
 use uint::FromStrRadixErr;
 
+use crate::chain::endpoint::ChainStatus;
 use crate::chain::handle::ChainHandle;
 use crate::chain::tracking::TrackedMsgs;
-use crate::chain::ChainStatus;
 use crate::error::Error;
 use crate::util::bigint::U256;
