+++ relayer/src/upgrade_chain.rs
@@ -18,9 +18,10 @@ use ibc_proto::cosmos::upgrade::v1beta1::Plan;
 use ibc_proto::google::protobuf::Any;
 use ibc_proto::ibc::core::client::v1::UpgradeProposal;
 
+use crate::chain::cosmos::CosmosSdkChain;
+use crate::chain::endpoint::ChainEndpoint;
 use crate::chain::requests::QueryClientStateRequest;
 use crate::chain::tracking::TrackedMsgs;
-use crate::chain::{ChainEndpoint, CosmosSdkChain};
 use crate::config::ChainConfig;
 use crate::error::Error;
